Fix gpt2-large quantization format from FP32 to FP16
Correct the quantization format for gpt2-large to match the actual
torch.float16 precision used by the memory optimization config.
All models use FP16 by default unless explicitly quantized to INT4.

+++ src/models/model_configs.py
@@ -41,7 +41,7 @@ AVAILABLE_MODELS = {
         top_p=0.9,
         estimated_memory_gb=1.5,  # Actual: 1.48GB from logs
         requires_system_prompt=False,
-        quantization_format="FP32"
+        quantization_format="FP16"
     ),
     
     "llama-3.2-1b": ModelConfig(
